comments: cut the judging pipeline down to its essentials
Drop ~300 lines of custom code the agent fan-out made redundant.

Apply re-extracts each judged file and matches verdicts to comments by id at
their current position, so the manifest, the recorded-range join, and the
text-drift check all go away. A comment that moved or changed gets a new id and
no longer matches, which is the drift check.

Move the oracle-only batch helpers (formatBatch, parseBatchVerdicts,
judgeComments) into evals/oracle.ts, leaving judge.ts as the shared prompt and
verdict-validation primitives.

Replace the hand-rolled unified-diff parser with parse-diff, and render the
oracle's model input as escaped XML via fast-xml-parser instead of a custom
banner format.

Delete the advisory tells layer. The judge's narration and section-divider
categories already cover it.
