+++ plugins/comments/evals/eval.ts
@@ -5,14 +5,15 @@ import { Glob } from "bun";
 import { cli } from "cleye";
 import { table } from "table";
 import type { CommentKind, Language } from "../detection/types";
+import { loadPrompt } from "../judge/judge";
+import { SLOP_CATEGORIES, type SlopCategory, type Verdict } from "../judge/schema";
 import {
+  anthropicCommentJudge,
   type CommentJudge,
   type CommentJudgeInput,
+  JUDGE_MODEL,
   judgeComments,
-  loadPrompt,
-} from "../judge/judge";
-import { SLOP_CATEGORIES, type SlopCategory, type Verdict } from "../judge/schema";
-import { anthropicCommentJudge, JUDGE_MODEL } from "./oracle";
+} from "./oracle";
 
 /**
  * One labeled comment with the surrounding context the judge sees. `label`
