+++ plugins/comments/package.json
@@ -4,10 +4,12 @@
   "type": "module",
   "dependencies": {
     "cleye": "^2.2.1",
+    "parse-diff": "^0.12.0",
     "shiki": "^4.3.0",
     "table": "^6.9.0"
   },
   "devDependencies": {
-    "@anthropic-ai/sdk": "^0.104.1"
+    "@anthropic-ai/sdk": "^0.104.1",
+    "fast-xml-parser": "^5.9.3"
   }
 }
